+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.2] - 2025-10-25
 - Describe PrimKind in opcodes.def
+- Buld in Windows
 
 ## [0.1.1] - 2025-10-25
 - Add lables to opcodes